% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attritevis_fns.R
\name{attrition}
\alias{attrition}
\title{Transform dataframe into an attrition dataframe.}
\usage{
attrition(data)
}
\arguments{
\item{data}{A data.frame where variables are ordered by survey questions, such that earlier survey questions appear in smaller valued columns.}
}
\value{
An attrition dataframe that contains the following variables:
 \describe{
    \item{attrited}{How many respondents attrited (left the survey) at each question}
    \item{proportion}{Number of attrited respondents / number of respondents who entered survey.}
    \item{prop_q}{Number of attrited respondents / number of respondents entering into the question}
    \item{questions}{Question names}
    \item{responded}{How many respondents responded in each question}
    \item{prop_r}{Number of respondents who responded / number of respondents who entered survey}
 }
}
\description{
The attrition dataframe indicates, per variable, how many respondents attrited. Note that this dataframe does not include "skippers", i.e. respondents who skipped questions. The dataframe also includes a variable that is the proportion of total N attrited, calculated as number of attrited respondents / starting N, and a variable that is the proportion of attrited by n at Qi, calculated as attrited respondents /  number of respondents entering into the question.
}
\examples{
attrition(empathy)
}
