% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attritevis_fns.R
\name{balance_attrite}
\alias{balance_attrite}
\title{Test for balance of attrition rate across specified treatment and control groups.}
\usage{
balance_attrite(data, treatment, question)
}
\arguments{
\item{data}{A data.frame from which \code{treatment} and \code{question} are taken.}

\item{treatment}{A character string that corresponds to the name of the treatment variable. Note that values of said variable must be specified as "treatment" and "control".}

\item{question}{A character string that corresponds to the name of the point in survey (question), for which balance test is required.}
}
\value{
A logistic regression, regressing attrition (remain in survey=0, attrited=1) over specified treatment.
}
\description{
Tests whether specified treatment causes attrition in a specified question.
}
\examples{
balance_attrite(data = empathy,
     treatment = "treat1",
     question = "Happy_3_1")
}
