% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attritevis_fns.R
\name{balance_cov}
\alias{balance_cov}
\title{Test for balance of covariates across specified treatment and control groups.}
\usage{
balance_cov(
  data,
  treatment,
  question,
  factor = FALSE,
  factor_name = NULL,
  p_adjust = NULL
)
}
\arguments{
\item{data}{A data.frame from which \code{treatment} and \code{question} are taken.}

\item{treatment}{A character string that corresponds to the name of the treatment variable. Note that values of said variable must be specified as "treatment" and "control".}

\item{question}{A character string that corresponds to the name of the point in survey (question), for which balance test is required.}

\item{factor}{Logical argument that specifies whether question is a factor. Default is \code{factor=  FALSE} (i.e. question is a numeric or integer).}

\item{factor_name}{Character that corresponds to specific factor (i.e. female), if question is a factor (i.e. sex).}

\item{p_adjust}{Vector of numbers that correspond to p-values obtained in all tests. Use this to adjust for p-values if running multiple tests.}
}
\value{
A t-test if covariate is a numeric or integer, and a 2-sample proportion test if covariate is a factor.
}
\description{
Tests whether specified covariates are balanced across specified treatment and control groups. Output is a t-test if covariate is a numeric or integer, and a 2-sample proportion test if covariate is a factor.
}
\examples{
balance_cov(data = empathy,
     treatment = "treat1",
     question = "sex",
     factor = TRUE,
     factor_name = "female")
}
