% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attritevis_fns.R
\name{plot_attrition}
\alias{plot_attrition}
\title{Plot attrition/response in survey data over time.}
\usage{
plot_attrition(
  data,
  y = "attrited",
  freq = TRUE,
  treatment_q = NULL,
  outcome_q = NULL,
  mycolors = NULL,
  title = NULL,
  total = TRUE,
  tline = TRUE,
  outcomeline = TRUE
)
}
\arguments{
\item{data}{A data.frame where variables are ordered by survey questions, such that earlier survey questions appear in smaller valued columns.}

\item{y}{A character that corresponds to the Y axis. When y = "attrited", attrition is plotted. When y= "responded" responses are plotted. Default is y = "attrited".}

\item{freq}{A logical argument that notes whether Y axis is a raw number or a proportion. Default is freq=TRUE, which is the frequency of attrited/responded respondents. When freq=FALSE Y axis is the proportion of total N (attrited/responded), calculated as number of attrited/responded divided by the number of respondents entering into the question.}

\item{treatment_q}{A character of name(s) of question(s) in which treatments were administered. Marked in the plot with a red vertical line.}

\item{outcome_q}{A character of name(s) of outcome question(s). Marked in the plot with a blue vertical line.}

\item{mycolors}{A character string of color names to be used as values in "scale_colour_manual" argument in ggplot. Default is mycolors=NULL, which defaults to grayscale. "mycolors" must be == length of the unique values of the "treatment_q" variable.}

\item{title}{A character string to be used for plot title.}

\item{total}{A logical argument that notes whether the total # of attrited/responded is plotted. Default is TRUE. Argument can be changed to FALSE only when "treatment_q" is full.}

\item{tline}{A logical argument that allows users to remove  treatment vline, default is tline=TRUE}

\item{outcomeline}{A logical argument that allows users to remove outcome vlines, default is outcomeline=TRUE}
}
\value{
An attrition plot.
}
\description{
The function allows us to plot attrition across survey questions, indicating where treatment and outcome questions were collected.
}
\examples{
# Simply plugging in the dataset into the function yields a figure
# that plots the number of respondents that attrited (left the survey
# completely) over each question in the study.
plot_attrition(empathy)

# When users specify freq=FALSE, the y axis plots the proportion of
# attrited.
plot_attrition(empathy,
     freq = FALSE)

# Users can further specify y="responded" to account for response,
# rather than attrition. This argument can be used with either freq=TRUE
# (default), or freq=FALSE, plotting response or proportion of responded,
# accordingly.
plot_attrition(empathy,
     y = "responded")

# Using the outcome_q argument, users can specify where outcome questions
# were measured. These are noted with gray vertical lines.
plot_attrition(empathy,
     outcome_q = c("cards1", "cards2", "cards3"))

# When treatment_q, which corresponds to treatment variable, is not NULL,
# the plot both notes where treatment was collected with a vertical line,
# and breaks down attrition by treatment conditions.
plot_attrition(empathy,
     y = "responded",
     outcome_q = c("cards1", "cards2", "cards3"),
     treatment_q = "treat1")
}
