% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attritevis_fns.R
\name{vis_miss_treat}
\alias{vis_miss_treat}
\title{Visualize attrition and missingness}
\usage{
vis_miss_treat(data, treatment_q = NULL)
}
\arguments{
\item{data}{A data.frame where variables are ordered by survey questions, such that earlier survey questions appear in smaller valued columns.}

\item{treatment_q}{A character string that corresponds to treatment variable. If \code{treatment_q = NULL}, missingness map appears for all data. If \code{treatment_q != NULL}, missingness is faceted by condition.}
}
\value{
A plot that shows pattern of missingness.
}
\description{
Calls the \code{vis_miss} function from \code{visdat} package. We allow users to facet missingness by conditions, creating several missingness maps per condition.
}
\examples{
vis_miss_treat(empathy)

# The function allows users to facet missingness by conditions, creating
# several missingness maps per condition, and marks treatment variable with
# a red vertical line.
vis_miss_treat(empathy,
     treatment_q = "treat1")
}
